Major changes - Enum course deprecation
Was about to add the whole Enumeration course's notes altogether, access got revoked prior to being able to complete that, made some minor changes to previous courses' notes as well.

Deprecated the enumeration course's notes to instead work on the new one by Alexis Ahmed for now.

Added some additional content that might not be present in the course, but is good to know about, MAC spoofing with Nmap for instance.